requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
bcrypt
fastapi-mail
pydantic
pydantic-settings